Handle the case in which the user schedules regular notifications
Because this version of the plugin supports categories and actions, it assumes
that they always exist, and breaks if they don't. Specifically, it breaks
because the array we use for permission checking has NSNull objects.

I am not sure why it returns NSNull objects instead of just being a shorter
array, but it is what it is. Instead of forcing the use of a category and
empty action list every time, we can just check and ignore NSNull objects.

This seems to work :)